feat: standalone classes for snapserver and snapclient processes
